refactor(web): rename overview to installation terminology
Aligns internal naming with user-facing terminology. "Overview" was
developer jargon from an earlier version when the main menu item was called
"Overview". Now the breadcrumb shows "Installation" and users see a
"summary" of settings, so the code should reflect that.

Changes:
- Renamed directory: components/overview -> components/installation
- Renamed component: OverviewPage -> InstallationPage
- Updated path constant: ROOT.overview removed, ROOT.installation renamed to
  ROOT.summary (more intuitive - matches what users conceptually see)
- Added legacy redirect: /overview -> /installation for bookmarked URLs
- Updated all references in router, breadcrumbs, and form submit links
- Updated comments and test expectations
