+++ web/src/components/installation/HostnameSummary.test.tsx
