+++ web/src/components/installation/HostnameSummary.tsx
