+++ web/src/components/installation/L10nSummary.test.tsx
