+++ web/src/components/installation/L10nSummary.tsx
