+++ web/src/components/installation/NetworkSummary.test.tsx
