+++ web/src/components/installation/NetworkSummary.tsx
