+++ web/src/components/installation/RegistrationSummary.test.tsx
