+++ web/src/components/installation/RegistrationSummary.tsx
