+++ web/src/components/installation/SoftwareSummary.test.tsx
