+++ web/src/components/installation/SoftwareSummary.tsx
