+++ web/src/components/installation/StorageSummary.test.tsx
