+++ web/src/components/installation/StorageSummary.tsx
