+++ web/src/components/installation/SystemInformationSection.test.tsx
