+++ web/src/components/installation/SystemInformationSection.tsx
