+++ web/src/components/installation/SystemSummary.test.tsx
