+++ web/src/components/installation/SystemSummary.tsx
