+++ web/src/routes/paths.ts
@@ -49,8 +49,7 @@ const REGISTRATION = {
 const ROOT = {
   root: "/",
   login: "/login",
-  overview: "/overview",
-  installation: "/installation",
+  summary: "/installation",
   installationProgress: "/installation/progress",
   installationFinished: "/installation/finished",
   installationExit: "/installation/exit",
